fix: preset cache poisoning, tidy errors, and GLX crash
- CMakePresets.json: every preset sharing build/ now explicitly sets all
  toggle options (ENABLE_SANITIZERS, ENABLE_PROFILING, ENABLE_CLANG_TIDY)
  so switching presets never leaves stale cache variables
- cmd_atlas.cc: cast to size_t before multiply to fix
  bugprone-implicit-widening-of-multiplication-result (unblocks game-tidy)
- sdl_init.cc: always use EGL instead of GLX — EGL is the modern
  platform-agnostic path, works on X11/Wayland/future targets, and
  avoids GLX breakage under ASan
- Add CMake and CTest improvements design doc

+++ src/cmd_atlas.cc
@@ -131,9 +131,9 @@ void CollectImages(const char* dir, const char* prefix, bool recursive,
 void BlitSprite(uint8_t* atlas, int atlas_w, const SpriteInput& sprite,
                 int dst_x, int dst_y) {
   for (int y = 0; y < sprite.height; ++y) {
-    const uint8_t* src = sprite.pixels + y * sprite.width * 4;
-    uint8_t* dst = atlas + ((dst_y + y) * atlas_w + dst_x) * 4;
-    memcpy(dst, src, sprite.width * 4);
+    const uint8_t* src = sprite.pixels + (size_t)y * sprite.width * 4;
+    uint8_t* dst = atlas + ((size_t)(dst_y + y) * atlas_w + dst_x) * 4;
+    memcpy(dst, src, (size_t)sprite.width * 4);
   }
 }
